requirements.txt
aiohttp==3.7.3
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
certifi==2020.11.8
chardet==3.0.4
Django==3.1.4
django-3-jet==1.0.8
django-admin-honeypot==1.1.0
django-admin-interface==0.14.2
django-appconf==1.0.4
django-blacklist==0.4.1
django-colorfield==0.4.1
django-cors-headers==3.5.0
django-countries==7.0
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-ipware==3.0.2
django-multi-captcha-admin==1.0.0
django-ratelimit==3.0.1
django-recaptcha==1.3.0
django-recaptcha2==1.4.1
django-tracking-analyzer==1.1.1
django-user-agents==0.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
geoip2==4.1.0
google-api-python-client==1.4.1
httplib2==0.18.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
maxminddb==2.0.3
mccabe==0.6.1
multidict==5.1.0
oauth2client==4.1.3
Pillow==8.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
PyJWT==2.0.0
pylint==2.6.0
pytz==2020.4
requests==2.25.0
rsa==4.7
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.2
user-agents==2.2.0
wrapt==1.12.1
yarl==1.6.3
